requirements.txt
fastapi
uvicorn[standard]

sqlmodel
psycopg2-binary

python-dotenv

passlib[bcrypt]
python-jose[cryptography]
email-validator

python-multipart
aiofiles

stripe
